Fix for some bad behavior in issue #10
We're seeing behavior described in #10 , and this change fixes the issue for us.

Actually what's happening for us is that by using `array_merge()` here, tokens that are not in `$body_subject` are simply not appearing in `$tokens`. Issue #10 doesn't describe the issue in exactly that way, but there's some chance that indeed that's the situation.